Added javadoc

+++ src/main/java/net/lenni0451/lambdaevents/EventHandler.java
@@ -5,12 +5,26 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+/**
+ * Mark a method or field as an event handler<br>
+ * The method can have no parameters or one parameter which is the event<br>
+ * The field has to be a {@link Runnable} or a {@link java.util.function.Consumer}<br>
+ * If no event parameter is specified (methods without parameter/consumer fields), {@link EventHandler#events()} has to be specified
+ */
 @Target({ElementType.METHOD, ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
 public @interface EventHandler {
 
-    byte priority() default 0;
+    /**
+     * The priority of the handler<br>
+     * higher = called first<br>
+     * lower = called last
+     */
+    int priority() default 0;
 
+    /**
+     * The events the handler listens to when no event parameter is present
+     */
     Class<?>[] events() default {};
 
 }
